Extract nox session to conftest.py
